requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
